Hoist shared constants interfaces to @glimmer/interfaces
Previously, there was an (undeclared in package.json) circular dependency between @glimmer/program and @glimmer/opcode-compiler, because the program was importing compile-time interfaces from the opcode compiler. This commit breaks the circular dependency by hoisting the shared interfaces to @glimmer/interfaces.
