+++ packages/@glimmer/opcode-compiler/lib/interfaces.ts
@@ -4,21 +4,11 @@ import {
   SymbolTable,
   Option,
   BlockSymbolTable,
-  Opcode,
-  ComponentCapabilities
+  ComponentCapabilities,
+  CompileTimeProgram
 } from '@glimmer/interfaces';
 import { Core, SerializedTemplateBlock } from '@glimmer/wire-format';
 import { Macros } from './syntax';
-
-export interface CompileTimeHeap {
-  push(name: /* TODO: Op */ number, op1?: number, op2?: number, op3?: number): void;
-  malloc(): VMHandle;
-  finishMalloc(handle: VMHandle, scopeSize: number): void;
-
-  // for debugging
-  getaddr(handle: VMHandle): number;
-  sizeof(handle: VMHandle): number;
-}
 
 export interface EagerResolver<Specifier> {
   getCapabilities(specifier: Specifier): ComponentCapabilities;
@@ -37,31 +27,7 @@ export interface CompilableTemplate<S extends SymbolTable> {
 
 export type CompilableBlock = CompilableTemplate<BlockSymbolTable>;
 
-export interface CompileTimeProgram {
-  [key: number]: never;
-
-  constants: CompileTimeConstants;
-  heap: CompileTimeHeap;
-
-  opcode(offset: number): Opcode;
-}
-
 export type Primitive = undefined | null | boolean | number | string;
-
-export interface CompileTimeConstants {
-  string(value: string): number;
-  stringArray(strings: string[]): number;
-  array(values: number[]): number;
-  table(t: SymbolTable): number;
-  handle(specifier: Opaque): number;
-  serializable(value: Opaque): number;
-  float(value: number): number;
-  negative(value: number): number;
-}
-
-export interface CompileTimeLazyConstants extends CompileTimeConstants {
-  other(value: Opaque): number;
-}
 
 export type ComponentArgs = [Core.Params, Core.Hash, Option<CompilableBlock>, Option<CompilableBlock>];
 export type Specifier = Opaque;
