+++ packages/@glimmer/opcode-compiler/lib/syntax.ts
@@ -1,10 +1,10 @@
-import { Option, Opaque } from '@glimmer/interfaces';
+import { CompileTimeProgram, Option, Opaque } from '@glimmer/interfaces';
 import { assert, dict, unwrap, EMPTY_ARRAY } from '@glimmer/util';
 import { Register } from '@glimmer/vm';
 import * as WireFormat from '@glimmer/wire-format';
 import * as ClientSide from './client-side';
 import OpcodeBuilder, { CompileTimeLookup, OpcodeBuilderConstructor } from "./opcode-builder";
-import { CompilableBlock, CompileTimeProgram } from './interfaces';
+import { CompilableBlock } from './interfaces';
 
 import Ops = WireFormat.Ops;
